Resized Input and submit button in Main-lower/ Set a space-around to all flex children of each main-section's direct child/
For the input button, the clamp method would be most optimal for making its width more responsive.

Typography needs to become responsive

List items in nav are to be inline and space-around should probably be used on them too.

Nav logo should also be placed correctly.

Images should also scale responsively

The bezier curve needs to be added as a background to the main-middle section

The footer also needs to worked on